[MOV] (*,)timesheet(,*): rename `hr_timesheet` to `account_analytic_line`
* This commit renames the `hr_timesheet` model and view files to
  `account_analytic_line`.
* We generally avoid functional naming for model or view files, and having the
  code for the `account.analytic.line` model inside a file named `hr_timesheet`
  does not add much clarity to devs.
