+++ addons/hr_timesheet/models/__init__.py
@@ -1,10 +1,10 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import account_analytic_line
 from . import account_analytic_line_calendar_employee
 from . import analytic_applicability
 from . import hr_employee
 from . import hr_employee_public
-from . import hr_timesheet
 from . import ir_http
 from . import ir_ui_menu
 from . import res_company
