+++ addons/sale_timesheet/__manifest__.py
@@ -17,11 +17,11 @@ have real delivered quantities in sales orders.
         'data/sale_service_data.xml',
         'security/ir.model.access.csv',
         'security/sale_timesheet_security.xml',
+        'views/account_analytic_line_views.xml',
         'views/account_invoice_views.xml',
         'views/sale_order_views.xml',
         'views/product_views.xml',
         'views/project_task_views.xml',
-        'views/hr_timesheet_views.xml',
         'views/res_config_settings_views.xml',
         'views/sale_timesheet_portal_templates.xml',
         'views/project_sharing_views.xml',
