+++ addons/sale_timesheet/report/timesheets_analysis_report.py
@@ -3,7 +3,7 @@
 
 from odoo import fields, models, api
 
-from odoo.addons.sale_timesheet.models.hr_timesheet import BILLABLE_TYPES
+from odoo.addons.sale_timesheet.models.account_analytic_line import BILLABLE_TYPES
 
 
 class TimesheetsAnalysisReport(models.Model):
