+++ addons/sale_timesheet/views/account_analytic_line_views.xml
